update to long to hold gigabyte

+++ LA3/LA3.h
@@ -7,7 +7,7 @@ void assignB(double* mat, int n, int m);
 void dotProduct(double* A, double* B, double* C, long n);
 double* mm(double* A, double* B, int n, int m, int p);
 void MPI_mm(double* A, double* B, double* C, int n, int m, int p, int rank, int numranks);
-void printMatrix(double* mat, int n, int m);
+void printMatrix(double* mat, long n, long m);
 void transposeMatrix(double* A, int m, int n);
 
 #endif
